feat: combine 2 services into 1 - authService,  use errors for validation instead of alerts, add loaders for authorization, add users collection, make notes subcollection of uer

+++ src/app/shared/services/data.service.ts
@@ -2,7 +2,7 @@ import { Injectable, Query } from '@angular/core';
 import { AngularFirestore, DocumentChangeAction } from '@angular/fire/compat/firestore';
 import { AngularFireStorage } from '@angular/fire/compat/storage';
 import { QueryDocumentSnapshot } from '@angular/fire/firestore';
-import { INote } from '@customTypes/models';
+import { INote, IUser } from '@customTypes/models';
 import { map, switchMap, take } from 'rxjs/operators';
 
 
@@ -13,20 +13,25 @@ export class DataService {
 
   constructor(private afs : AngularFirestore, private fireStorage : AngularFireStorage) { }
 
-  addNote(note : INote) {
+  addNote(userId: string, note : INote) {
     note.id = this.afs.createId();
-    return this.afs.collection('/notes').add(note);
+    return this.afs.collection(`/users/${userId}/notes`).add(note);
   }
 
-  getNotesCount() {
-    return this.afs.collection('/notes').get().pipe(
+  addUser(user: IUser) {
+    user.id = this.afs.createId()
+    return this.afs.collection('/users').add(user)
+  }
+
+  getNotesCount(userId: string) {
+    return this.afs.collection(`/users/${userId}/notes`).get().pipe(
       map(snapshot => snapshot.size)
     );
   }
 
-  getAllNotes() {
+  getAllNotes(userId: string) {
     // return this.afs.collection('/notes').snapshotChanges();
-    return this.afs.collection('/notes', ref => ref.orderBy('body.time', 'desc')).snapshotChanges()
+    return this.afs.collection(`/users/${userId}/notes`, ref => ref.orderBy('body.time', 'desc')).snapshotChanges()
   }
 
   // getLimitedNotes(startTimeStamp: number | undefined, limit: number) {
@@ -36,19 +41,19 @@ export class DataService {
   //   return this.afs.collection('/notes', ref => ref.orderBy('body.time', 'desc').limit(limit)).snapshotChanges()
   // }
 
-  getLimitedNotes(lastNoteId: string | undefined, limit: number) {
-    let query = this.afs.collection('/notes', ref => ref.orderBy('body.time', 'desc'));
+  getLimitedNotes(userId: string, lastNoteId: string | undefined, limit: number) {
+    // let query = this.afs.collection('/notes', ref => ref.orderBy('body.time', 'desc'));
 
     if (lastNoteId) {
       // Получаем документ с переданным startTimeStamp
-      return this.afs.doc(`/notes/${lastNoteId}`).get().pipe(
+      return this.afs.doc(`/users/${userId}/notes/${lastNoteId}`).get().pipe(
         take(1),
         switchMap((doc: any) => {
           // Приводим тип данных к ожидаемому типу QueryDocumentSnapshot<INote>
           const queryDoc = doc as QueryDocumentSnapshot<INote>;
           // Получаем документы, идущие после переданного документа
           // return query.ref.startAfter(queryDoc).limit(limit).get();
-          return this.afs.collection('/notes', ref => ref.orderBy('body.time', 'desc').startAfter(queryDoc).limit(limit)).get();
+          return this.afs.collection(`/users/${userId}/notes`, ref => ref.orderBy('body.time', 'desc').startAfter(queryDoc).limit(limit)).get();
         }),
         map(snapshot => {
           return snapshot.docs.map((doc: any) => {
@@ -60,7 +65,7 @@ export class DataService {
       );
     }
 
-    return this.afs.collection('/notes', ref => ref.orderBy('body.time', 'desc').limit(limit)).get().pipe(
+    return this.afs.collection(`/users/${userId}/notes`, ref => ref.orderBy('body.time', 'desc').limit(limit)).get().pipe(
       map(snapshot => {
         return snapshot.docs.map((doc: any) => {
           const data = doc.data();
@@ -85,12 +90,12 @@ export class DataService {
   //     );
   // }
 
-  deleteNote(note : INote) {
-     this.afs.doc('/notes/'+note.id).delete();
+  deleteNote(userId: string, note : INote) {
+     this.afs.doc(`/users/${userId}/notes/${note.id}`).delete();
   }
 
-  updateNote(id : string, newNote: INote) {
+  updateNote(userId: string, noteId : string, newNote: INote) {
     console.log('update')
-    return this.afs.doc('/notes/'+id).update(newNote)
+    return this.afs.doc(`/users/${userId}/notes/${noteId}`).update(newNote)
   }
 }
